more ui changes, changed logic for setting alarms and calculating times, hide pseudo-fab on saved fragment when inputing name, name input in saved fragment now updates database upon completion

+++ app/src/main/java/com/example/crumb/Dao/ScheduleDao.kt
@@ -53,7 +53,7 @@ interface ScheduleDao {
 
     // start time
     @Query("UPDATE schedule_table SET start_time = :start where id = :id")
-    suspend fun updateStartTime(id : Long, start : Int)
+    suspend fun updateStartTime(id : Long, start : Long)
     // end time
     @Query("UPDATE schedule_table SET end_time = :end where id = :id")
     suspend fun updateEndTime(id : Long, end : Int)
